add docs for all pub stuff
